""" Add SmartDataHandler class for reading and writing data to a pickle file
This commit adds a new module named `data_handler.py`, which contains a class
named `SmartDataHandler`. This class has two methods: `read_data()` and
`write_data()`, which can be used to read and write data to a file named
`smart_data.pkl` using the `pickle` module.

The `read_data()` method reads the contents of the file and returns the resulting
dictionary, while the `write_data()` method overwrites the contents of the file
with the given dictionary.

This module can be used as a reusable component in other Python projects that
need to store and retrieve data in a simple file format.
"""